# src/openapi.yaml
openapi: 3.0.0
info:
  title: Boys State App API
  description: >
    **Disclaimer:** This project is being developed to support Boys State &
    Girls State programs affiliated with the American Legion, but is not
    created, funded, or officially supported by the American Legion. No
    endorsement or sponsorship is implied. All branding, configuration, and
    operational decisions are made independently by the app’s creators and
    participating programs.


    The API provides authentication endpoints and health checks used by the Boys
    State App mobile and web clients. The following documentation describes the
    requests and responses in detail.
  version: 1.0.0
servers:
  - url: https://boysstateappservices.up.railway.app
...
paths:
...
  /register:
    post:
      tags:
        - auth
      summary: Register a new user
      description: >
        Creates a new user account using an email and password. The email must
        be unique. A successful request returns a confirmation message.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: User's email address
                  example: user@example.com
                password:
                  type: string
                  description: Plain text password to register
                  example: mySecret
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: User created
              example:
                message: User created
        "400":
          description: Email already exists or invalid input
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: User already exists
              example:
                error: User already exists
  /login:
    post:
      tags:
        - auth
      summary: Authenticate an existing user
      description: >
        Checks the provided email and password and returns a success message
        when the credentials are valid.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  description: Registered email address
                  example: user@example.com
                password:
                  type: string
                  description: Account password
                  example: mySecret
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT authentication token containing user id and email (expires
                      after 30 minutes)
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MT
                      Y4NTg0OTc1NX0.LXrlo3wrX7rQiGZnYuxRKgXUxRvQPgFP5yxv1rdqe8E
              example:
                token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MT
                  Y4NTg0OTc1NX0.LXrlo3wrX7rQiGZnYuxRKgXUxRvQPgFP5yxv1rdqe8E
        "400":
          description: Missing credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Email and password required
              example:
                error: Email and password required
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Invalid credentials
              example:
                error: Invalid credentials
        "429":
          description: Too many failed login attempts
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Too many login attempts
              example:
                error: Too many login attempts
...
  /api/programs/{programId}/application:
    get:
      tags:
        - applications
      summary: Get application configuration
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Application configuration
          content:
            application/json:
              schema:
                type: object
              example:
                applicationId: app123
                title: Delegate Application
                description: Please complete the application
                questions:
                  - id: 1
                    order: 1
                    type: text
                    text: First name
                    required: true
                    options: []
                    fields: []
        "404":
          description: Not found
          content:
            application/json:
              example: {}
    post:
      tags:
        - applications
      summary: Create application configuration
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created configuration
          content:
            application/json:
              schema:
                type: object
              example:
                applicationId: app123
                title: Delegate Application
                description: Please complete the application
                questions:
                  - id: 1
                    order: 1
                    type: text
                    text: First name
                    required: true
                    options: []
                    fields: []
        "403":
          description: Forbidden
          content:
            application/json:
              example: {}
        "404":
          description: Not found
          content:
            application/json:
              example: {}
      security:
        - bearerAuth: []
    put:
      tags:
        - applications
      summary: Update application configuration
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created configuration
          content:
            application/json:
              schema:
                type: object
              example:
                applicationId: app123
                title: Delegate Application
                description: Please complete the application
                questions:
                  - id: 1
                    order: 1
                    type: text
                    text: First name
                    required: true
                    options: []
                    fields: []
        "403":
          description: Forbidden
          content:
            application/json:
              example: {}
        "404":
          description: Not found
          content:
            application/json:
              example: {}
      security:
        - bearerAuth: []
    delete:
      tags:
        - applications
      summary: Delete application configuration
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              example: {}
        "403":
          description: Forbidden
          content:
            application/json:
              example: {}
        "404":
          description: Not found
          content:
            application/json:
              example: {}
      security:
        - bearerAuth: []
  /api/programs/{programId}/application/responses:
    post:
      tags:
        - applications
      summary: Submit application response
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    properties:
                      questionId:
                        type: integer
                      value:
                        type: string
            example:
              answers:
                - questionId: 1
                  value: John Doe
      responses:
        "201":
          description: Created response
          content:
            application/json:
              example:
                responseId: resp123
        "404":
          description: Not found
          content:
            application/json:
              example: {}
    get:
      tags:
        - applications
      summary: List application responses
      parameters:
        - in: path
          name: programId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Responses
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
              example:
                - id: resp123
                  answers:
                    - questionId: 1
                      value: John Doe
        "403":
          description: Forbidden
          content:
            application/json:
              example: {}
        "404":
          description: Not found
          content:
            application/json:
              example: {}
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT